Full enchilada: SO3 is now SO<3>
