+++ gtsam/geometry/SO3.cpp
@@ -18,22 +18,23 @@
  * @date    December 2014
  */
 
-#include <gtsam/geometry/SO3.h>
 #include <gtsam/base/concepts.h>
+#include <gtsam/geometry/SO3.h>
 
 #include <Eigen/SVD>
 
 #include <cmath>
-#include <limits>
 #include <iostream>
+#include <limits>
 
 namespace gtsam {
 
-/* ************************************************************************* */
+//******************************************************************************
 namespace so3 {
 
-Matrix99 Dcompose(const SO3& R) {
+Matrix99 Dcompose(const SO3& Q) {
   Matrix99 H;
+  auto R = Q.matrix();
   H << I_3x3 * R(0, 0), I_3x3 * R(1, 0), I_3x3 * R(2, 0),  //
       I_3x3 * R(0, 1), I_3x3 * R(1, 1), I_3x3 * R(2, 1),   //
       I_3x3 * R(0, 2), I_3x3 * R(1, 2), I_3x3 * R(2, 2);
@@ -47,7 +48,8 @@ Matrix3 compose(const Matrix3& M, const SO3& R, OptionalJacobian<9, 9> H) {
 }
 
 void ExpmapFunctor::init(bool nearZeroApprox) {
-  nearZero = nearZeroApprox || (theta2 <= std::numeric_limits<double>::epsilon());
+  nearZero =
+      nearZeroApprox || (theta2 <= std::numeric_limits<double>::epsilon());
   if (!nearZero) {
     sin_theta = std::sin(theta);
     const double s2 = std::sin(theta / 2.0);
@@ -79,9 +81,9 @@ ExpmapFunctor::ExpmapFunctor(const Vector3& axis, double angle, bool nearZeroApp
 
 SO3 ExpmapFunctor::expmap() const {
   if (nearZero)
-    return I_3x3 + W;
+    return SO3(I_3x3 + W);
   else
-    return I_3x3 + sin_theta * K + one_minus_cos * KK;
+    return SO3(I_3x3 + sin_theta * K + one_minus_cos * KK);
 }
 
 DexpFunctor::DexpFunctor(const Vector3& omega, bool nearZeroApprox)
@@ -121,7 +123,7 @@ Vector3 DexpFunctor::applyInvDexp(const Vector3& v, OptionalJacobian<3, 3> H1,
   if (H1) {
     Matrix3 D_dexpv_omega;
     applyDexp(c, D_dexpv_omega);  // get derivative H of forward mapping
-    *H1 = -invDexp* D_dexpv_omega;
+    *H1 = -invDexp * D_dexpv_omega;
   }
   if (H2) *H2 = invDexp;
   return c;
@@ -129,72 +131,117 @@ Vector3 DexpFunctor::applyInvDexp(const Vector3& v, OptionalJacobian<3, 3> H1,
 
 }  // namespace so3
 
-/* ************************************************************************* */
+//******************************************************************************
+template <>
 SO3 SO3::AxisAngle(const Vector3& axis, double theta) {
   return so3::ExpmapFunctor(axis, theta).expmap();
 }
 
-/* ************************************************************************* */
+//******************************************************************************
+template <>
 SO3 SO3::ClosestTo(const Matrix3& M) {
   Eigen::JacobiSVD<Matrix3> svd(M, Eigen::ComputeFullU | Eigen::ComputeFullV);
   const auto& U = svd.matrixU();
   const auto& V = svd.matrixV();
   const double det = (U * V.transpose()).determinant();
-  return U * Vector3(1, 1, det).asDiagonal() * V.transpose();
+  return SO3(U * Vector3(1, 1, det).asDiagonal() * V.transpose());
 }
 
-/* ************************************************************************* */
+//******************************************************************************
+template <>
 SO3 SO3::ChordalMean(const std::vector<SO3>& rotations) {
-  //  See Hartley13ijcv:
-  //  Cost function C(R) = \sum sqr(|R-R_i|_F)
+  // See Hartley13ijcv:
+  // Cost function C(R) = \sum sqr(|R-R_i|_F)
   // Closed form solution = ClosestTo(C_e), where C_e = \sum R_i !!!!
-  Matrix3 C_e {Z_3x3};
+  Matrix3 C_e{Z_3x3};
   for (const auto& R_i : rotations) {
-    C_e += R_i;
+    C_e += R_i.matrix();
   }
   return ClosestTo(C_e);
 }
 
-/* ************************************************************************* */
-void SO3::print(const std::string& s) const {
-   std::cout << s << *this << std::endl;
- }
-
 //******************************************************************************
- Matrix3 SO3::Hat(const Vector3& xi) { return skewSymmetric(xi); }
-
- /* ************************************************************************* */
- Vector3 SO3::Vee(const Matrix3& X) {
-   Vector3 xi;
-   xi(0) = -X(1, 2);
-   xi(1) = X(0, 2);
-   xi(2) = -X(0, 1);
-   return xi;
+template <>
+Matrix3 SO3::Hat(const Vector3& xi) {
+  // skew symmetric matrix X = xi^
+  Matrix3 Y = Z_3x3;
+  Y(0, 1) = -xi(2);
+  Y(0, 2) = +xi(1);
+  Y(1, 2) = -xi(0);
+  return Y - Y.transpose();
 }
 
-/* ************************************************************************* */
+//******************************************************************************
+template <>
+Vector3 SO3::Vee(const Matrix3& X) {
+  Vector3 xi;
+  xi(0) = -X(1, 2);
+  xi(1) = +X(0, 2);
+  xi(2) = -X(0, 1);
+  return xi;
+}
+
+//******************************************************************************
+template <>
+Matrix3 SO3::AdjointMap() const {
+  return matrix_;
+}
+
+//******************************************************************************
+template <>
 SO3 SO3::Expmap(const Vector3& omega, ChartJacobian H) {
   if (H) {
     so3::DexpFunctor impl(omega);
     *H = impl.dexp();
     return impl.expmap();
-  } else
+  } else {
     return so3::ExpmapFunctor(omega).expmap();
+  }
 }
 
+template <>
 Matrix3 SO3::ExpmapDerivative(const Vector3& omega) {
   return so3::DexpFunctor(omega).dexp();
 }
 
-/* ************************************************************************* */
-Vector3 SO3::Logmap(const SO3& R, ChartJacobian H) {
-  using std::sqrt;
+//******************************************************************************
+/* Right Jacobian for Log map in SO(3) - equation (10.86) and following
+ equations in G.S. Chirikjian, "Stochastic Models, Information Theory, and Lie
+ Groups", Volume 2, 2008.
+
+   logmap( Rhat * expmap(omega) ) \approx logmap(Rhat) + Jrinv * omega
+
+ where Jrinv = LogmapDerivative(omega). This maps a perturbation on the
+ manifold (expmap(omega)) to a perturbation in the tangent space (Jrinv *
+ omega)
+ */
+template <>
+Matrix3 SO3::LogmapDerivative(const Vector3& omega) {
+  using std::cos;
   using std::sin;
 
+  double theta2 = omega.dot(omega);
+  if (theta2 <= std::numeric_limits<double>::epsilon()) return I_3x3;
+  double theta = std::sqrt(theta2);  // rotation angle
+
+  // element of Lie algebra so(3): W = omega^
+  const Matrix3 W = Hat(omega);
+  return I_3x3 + 0.5 * W +
+         (1 / (theta * theta) - (1 + cos(theta)) / (2 * theta * sin(theta))) *
+             W * W;
+}
+
+//******************************************************************************
+template <>
+Vector3 SO3::Logmap(const SO3& Q, ChartJacobian H) {
+  using std::sin;
+  using std::sqrt;
+
   // note switch to base 1
-  const double& R11 = R(0, 0), R12 = R(0, 1), R13 = R(0, 2);
-  const double& R21 = R(1, 0), R22 = R(1, 1), R23 = R(1, 2);
-  const double& R31 = R(2, 0), R32 = R(2, 1), R33 = R(2, 2);
+  const Matrix3& R = Q.matrix();
+  const double &R11 = R(0, 0), R12 = R(0, 1), R13 = R(0, 2);
+  const double &R21 = R(1, 0), R22 = R(1, 1), R23 = R(1, 2);
+  const double &R31 = R(2, 0), R32 = R(2, 1), R33 = R(2, 2);
 
   // Get trace(R)
   const double tr = R.trace();
@@ -213,7 +260,7 @@ Vector3 SO3::Logmap(const SO3& R, ChartJacobian H) {
       omega = (M_PI / sqrt(2.0 + 2.0 * R11)) * Vector3(1.0 + R11, R21, R31);
   } else {
     double magnitude;
-    const double tr_3 = tr - 3.0; // always negative
+    const double tr_3 = tr - 3.0;  // always negative
     if (tr_3 < -1e-7) {
       double theta = acos((tr - 1.0) / 2.0);
       magnitude = theta / (2.0 * sin(theta));
@@ -225,53 +272,49 @@ Vector3 SO3::Logmap(const SO3& R, ChartJacobian H) {
     omega = magnitude * Vector3(R32 - R23, R13 - R31, R21 - R12);
   }
 
-  if(H) *H = LogmapDerivative(omega);
+  if (H) *H = LogmapDerivative(omega);
   return omega;
 }
 
-/* ************************************************************************* */
-Matrix3 SO3::LogmapDerivative(const Vector3& omega) {
-  using std::cos;
-  using std::sin;
+//******************************************************************************
+// Chart at origin for SO3 is *not* Cayley but actual Expmap/Logmap
 
-  double theta2 = omega.dot(omega);
-  if (theta2 <= std::numeric_limits<double>::epsilon()) return I_3x3;
-  double theta = std::sqrt(theta2);  // rotation angle
-  /** Right Jacobian for Log map in SO(3) - equation (10.86) and following equations in
-   * G.S. Chirikjian, "Stochastic Models, Information Theory, and Lie Groups", Volume 2, 2008.
-   * logmap( Rhat * expmap(omega) ) \approx logmap( Rhat ) + Jrinv * omega
-   * where Jrinv = LogmapDerivative(omega);
-   * This maps a perturbation on the manifold (expmap(omega))
-   * to a perturbation in the tangent space (Jrinv * omega)
-   */
-  const Matrix3 W = skewSymmetric(omega); // element of Lie algebra so(3): W = omega^
-  return I_3x3 + 0.5 * W +
-         (1 / (theta * theta) - (1 + cos(theta)) / (2 * theta * sin(theta))) *
-             W * W;
+template <>
+SO3 SO3::ChartAtOrigin::Retract(const Vector3& omega, ChartJacobian H) {
+  return Expmap(omega, H);
 }
 
-/* ************************************************************************* */
-static Vector9 vec(const SO3& R) { return Eigen::Map<const Vector9>(R.data()); }
+template <>
+Vector3 SO3::ChartAtOrigin::Local(const SO3& R, ChartJacobian H) {
+  return Logmap(R, H);
+}
 
-static const std::vector<const Matrix3> G({SO3::Hat(Vector3::Unit(0)),
-                                           SO3::Hat(Vector3::Unit(1)),
-                                           SO3::Hat(Vector3::Unit(2))});
+//******************************************************************************
+// local vectorize
+static Vector9 vec3(const Matrix3& R) {
+  return Eigen::Map<const Vector9>(R.data());
+}
 
-static const Matrix93 P =
-    (Matrix93() << vec(G[0]), vec(G[1]), vec(G[2])).finished();
+// so<3> generators
+static const std::vector<const Matrix3> G3({SO3::Hat(Vector3::Unit(0)),
+                                            SO3::Hat(Vector3::Unit(1)),
+                                            SO3::Hat(Vector3::Unit(2))});
 
-/* ************************************************************************* */
+// vectorized generators
+static const Matrix93 P3 =
+    (Matrix93() << vec3(G3[0]), vec3(G3[1]), vec3(G3[2])).finished();
+
+//******************************************************************************
+template <>
 Vector9 SO3::vec(OptionalJacobian<9, 3> H) const {
-  const SO3& R = *this;
+  const Matrix3& R = matrix_;
   if (H) {
-    // As Luca calculated (for SO4), this is (I3 \oplus R) * P
-    *H << R * P.block<3, 3>(0, 0), R * P.block<3, 3>(3, 0),
-        R * P.block<3, 3>(6, 0);
+    // As Luca calculated (for SO4), this is (I3 \oplus R) * P3
+    *H << R * P3.block<3, 3>(0, 0), R * P3.block<3, 3>(3, 0),
+        R * P3.block<3, 3>(6, 0);
   }
-  return gtsam::vec(R);
-};
+  return gtsam::vec3(R);
+}
+//******************************************************************************
 
-/* ************************************************************************* */
-
-} // end namespace gtsam
-
+}  // end namespace gtsam
